pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Takao Mizuno <takao.mizuno110@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.62.0"
sqlalchemy = "^1.3.20"
uvicorn = "^0.12.3"
python-decouple = "^3.3"
PyMySQL = "^0.10.1"
sqlalchemy_utils = "^0.36.8"
ERAlchemy = "^1.2.10"
aiofiles = "^0.6.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = "^1.7.4"
Django = "^3.1.7"
bpython = "^0.21"
pydantic = {extras = ["email"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
sqlalchemy-stubs = "^0.3"
flake8 = "^3.8.4"
autoflake = "^1.4"
isort = "^5.6.4"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
